restructured components into their own files
