+++ src/components/app.js
@@ -1,0 +1,16 @@
+import React, {Fragment, Component} from 'react';
+
+import Header from './header.js'
+import Main from './main.js'
+
+
+function App() {
+    return (
+        <Fragment>
+            <Header />
+            <Main />
+        </Fragment>
+    );
+}
+
+export default App;
