+++ src/components/header.js
@@ -1,0 +1,14 @@
+import React, {Fragment, Component} from 'react'
+
+class Header extends Component {
+    render() {
+      return (
+      <Fragment>
+        <h1>City Explorer</h1>
+        <p>Enter a location below to learn about the weather, events, restaurants, movies filmed there, and more!</p>
+      </Fragment>
+      );
+    }
+  }
+
+export default Header
