+++ src/components/map.js
@@ -1,0 +1,13 @@
+import React, {Fragment, Component} from 'react'
+
+class Map extends Component {
+    render() {
+      return (
+        <Fragment>
+          <iframe />
+        </Fragment>
+      );
+    }
+  }
+
+export default Map
