+++ src/components/result.js
@@ -1,0 +1,12 @@
+import React, {Fragment, Component} from 'react'
+
+
+class Result extends Component {
+    render() {
+        return (
+            <section>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed purus turpis, ultrices nec luctus ut, congue at nisi. Praesent commodo, lectus sed fermentum tempor, sapien nisi pretium odio, quis efficitur dui elit ut ligula. Orci varius natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Morbi consectetur tincidunt neque, ut aliquet libero aliquet sed.</section>
+        );
+    }
+}
+
+export default Result
